Move some types around
